// source/Schema/ServerStats.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ServerStats {
  Server: {
    game: string;
    version: string;
    name: string;
    mapName: string;
    dayTime: number;
    mapOverviewFilename: string;
    mapSize: number;
    Slots: {
      capacity: number;
      numUsed: number;
      Player: {
        isUsed: boolean;
        isAdmin: boolean;
        uptime: number;
        x: number;
        y: number;
        z: number;
        _text: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    Vehicles: {
      Vehicle: {
        name: string;
        category: string;
        type: string;
        x: number;
        y: number;
        z: number;
        fillTypes: string;
        fillLevels: number;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    Mods: {
      Mod: {
        name: string;
        author: string;
        version: string;
        hash: string;
        _text: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    Farmlands: {
      Farmland: {
        name: string;
        id: number;
        owner: number;
        area: number;
        x: number;
        z: number;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    Fields: {
      Field: {
        id: number;
        x: number;
        z: number;
        isOwned: boolean;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}